Adding script

+++ Registry/Restore-ClassicWindowsExplorerContextMenu.ps1
@@ -4,6 +4,11 @@ param(
     [Parameter(Mandatory = $false)][System.Management.Automation.SwitchParameter]$RestoreDefault
 )
 
+# Check if windows version is greater or equal to 11
+if ([System.Environment]::OSVersion.Version.Build -lt 22000) {
+    return "Windows version is lower than 11!"
+}
+
 if ($RestoreDefault) {
     Remove-Item -Path "HKCU:\Software\Classes\CLSID\{86ca1aa0-34aa-4e8b-a509-50c905bae2a2}\InprocServer32" -Force
     Remove-Item -Path "HKCU:\Software\Classes\CLSID\{86ca1aa0-34aa-4e8b-a509-50c905bae2a2}" -Force
